backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.2.1
plotly==5.19.0
scikit-learn==1.4.0
sqlalchemy==1.4.51
python-multipart==0.0.9
python-dotenv==1.0.1
joblib==1.3.2
numpy==1.26.4